% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{acfd}
\alias{acfd}
\title{acfd}
\usage{
acfd(dfs, lmin, lmax)
}
\arguments{
\item{dfs}{Uncorrected D values}

\item{lmin}{whatever lmin is, it wasnt well documented}

\item{lmax}{whatever lmax is, it wasnt well documented}
}
\description{
With the name of this function what this does your guess is as good as mine
Used for correcting d values
Take rawD and convert to the corrected D
}
