% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWSAnalysis.R
\name{AdaptPWSData}
\alias{AdaptPWSData}
\title{AdaptPWSData}
\usage{
AdaptPWSData(
  df,
  ExpName = "ExpName",
  ExpDate = "ExpDate",
  IsolateTimeStrat,
  IsolateCondStrat
)
}
\arguments{
\item{df}{a table produced by the PWS system}

\item{ExpName}{Name for the experiment}

\item{ExpDate}{Date of the experiment}

\item{IsolateTimeStrat}{A function to isolate the time variable from the
Path column}

\item{IsolateCondStrat}{A function to isolate the condition variable from the path column}
}
\description{
AdaptPWSData
}
