% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{ComputeB_SD}
\alias{ComputeB_SD}
\title{ComputeB_SD
helper function used to Compute the ACF used in the Size Distribution Model
This function evaluates the model of the Autocorrelation Function (ACF) of chromatin
for given input parameters. It is based on the methodology described in the pape
Characterizing chromatin packing scaling in whole nuclei using interferometric microscopy.}
\usage{
ComputeB_SD(d, lmin, lmax, r)
}
\arguments{
\item{d}{Numeric. The uncorrected fractal dimension (Db). Must be a scalar.}

\item{lmin}{Numeric. The minimum size in nanometers that the ACF is expected to be valid for. Must be a scalar.}

\item{lmax}{Numeric. The maximum size in nanometers that the chromatin is expected to be fractal for. Must be a scalar.}

\item{r}{Numeric. The size or sizes at which to evaluate the function. Can be a scalar or a 1D array.}
}
\value{
Numeric. Returns the computed value of the ACF (bnr) based on the input parameters.
The returned value will have the same length as the input r.
}
\description{
ComputeB_SD
helper function used to Compute the ACF used in the Size Distribution Model
This function evaluates the model of the Autocorrelation Function (ACF) of chromatin
for given input parameters. It is based on the methodology described in the pape
Characterizing chromatin packing scaling in whole nuclei using interferometric microscopy.
}
\examples{
d_example <- 2.5
lmin_example <- 10
lmax_example <- 200
r_example <- seq(1, 100, by = 1)
ComputeB_SD(d_example, lmin_example, lmax_example, r_example)

}
