% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{ConvertRMStoD}
\alias{ConvertRMStoD}
\title{This file contains an R adaptation of the RMS (root mean squared of reflectance across wavelengths
captured by PWS) to fractal dimension (D) code written in matlab by Aya Eid (refer to https://pubmed.ncbi.nlm.nih.gov/32870863) 
and later to my deduction refactored by Nick Antony. This implementation of that code is designed to
be a trimmed down implementation and is missing some features of the original code that I did
not require for my own research.}
\usage{
ConvertRMStoD(
  RMSvals,
  noiseRMS,
  phi = 0.35,
  Nf = 5e+05,
  ncCenterLambda = 585,
  na_i = 0.52,
  na_c = 1.49,
  IsCellGlassInterface = T,
  IsImmersedInOil = T,
  thickness = 2
)
}
\arguments{
\item{RMSvals}{a vector of RMS values}

\item{noiseRMS}{the amount of background noise (RMS values of the background image)}

\item{phi}{Not sure, using defaults in the MatLab script}

\item{Nf}{size of the packing domain}

\item{ncCenterLambda}{a constant taken from the script}

\item{na_i}{Illumination NA}

\item{na_c}{Collection NA}

\item{IsCellGlassInterface}{whether there is a cell glass interface, if true then the fresnel coefficients will reflect}

\item{IsImmersedInOil}{True/False if so the immersion_ri will be set to the RI of the glass found in ri_def}

\item{thickness}{thickness of material in uM}

\item{ImmersionRI}{The refractive index of the immersion media used for the objective}
}
\value{
a vector of D values
}
\description{
ConvertRMStoD
}
\details{
The main function of this software module. It converts RMS to D (Fractal Dimension values)
}
