% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{createRIFromGladstoneDale}
\alias{createRIFromGladstoneDale}
\title{Create an RI Definition Object Using the GladStoneGale Equation}
\usage{
createRIFromGladstoneDale(ri_media = 1.337, phi = 0.35)
}
\arguments{
\item{ri_media}{refractive index of the media}

\item{phi}{Phi the CVC of the nucleus.}
}
\value{
a named list with the following values
ri_chromatin: refractive index of chromatin
ri_glass:     refractive index of glass
ri_media:     refractive index of media
sigma_n:      SD of RI fluctutations in chromatin
}
\description{
The Refractive Index (RI object) is an object that contains RI information on 
assorted materials required for D calculations. The "object" also calculates values
for other constants which are calculated from tehse values.
Default values were pulled from
scripts in the original matlab code but I may not have fully understood their intentions
}
\details{
Original in matlab is an object but again the object use is pretty sparse and 
functions primarily as a named list so we will use that structure here is awell.
}
