% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{gamma_inc}
\alias{gamma_inc}
\title{Upper Incomplete Gamma Function}
\usage{
gamma_inc(x, a)
}
\arguments{
\item{x}{Numeric vector; function argument which must be non-negative.}

\item{a}{Numeric scalar; parameter of the incomplete gamma function, can be negative.}
}
\value{
Numeric vector of the same length as \code{x}. 
        Returns the computed values of the upper incomplete gamma function.
        NaN values are returned where elements of \code{x} are negative.
}
\description{
This function evaluates the upper incomplete gamma function \eqn{\Gamma(a, x)}
for non-negative values of the argument \code{x}. It extends the computation
to include negative values of the parameter \code{a}, which is not supported
by the standard gamma functions in R.
}
\details{
The function is an extension of the standard incomplete gamma
         functions, allowing for negative values of the parameter \code{a}.
         For \code{a > 0}, it uses the \code{pgamma} function from base R. 
         For \code{a = 0}, it relies on the exponential integral function
         from the \code{expint} package. For \code{a < 0}, the function 
         uses recursion to compute the values.
}
\examples{
x <- seq(0.01, 8, by = 0.01)
a_pos <- 1
a_neg <- -2.3
plot(x, gamma_incomplete(x, a_pos), type = 'l', col = 'blue',
     main = 'Upper Incomplete Gamma Function',
     xlab = 'x', ylab = expression(Gamma(a, x)))
lines(x, gamma_incomplete(x, a_neg), col = 'red')

}
