% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{LMaxCorrection}
\alias{LMaxCorrection}
\title{The function for correcting LMAX values}
\usage{
LMaxCorrection(dfs, lmin, Nf)
}
\arguments{
\item{dfs}{uncorrected D values}

\item{lmin}{not sure for now}

\item{Nf}{genomic length of the packing domain}
}
\description{
The function for correcting LMAX values
}
