% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{numericalInversion}
\alias{numericalInversion}
\title{numericalInversion}
\usage{
numericalInversion(df, xVals, Nf)
}
\arguments{
\item{xVals}{an array of x values to pass to func}

\item{func}{a function that accepts an array of values and returns y values}

\item{y}{the input to the inverted function for which you want x values}
}
\description{
for a function y = f(x) use interpolation and a lookup table to approximate
x = g(y). The function this function is supposed to use CalcNfs is hardcoded into this function
for simplicity. The interpolation method is slightly different between R and MatLab
and may cause some small variations
}
