% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWSAnalysis.R
\name{Plot_SampleWise}
\alias{Plot_SampleWise}
\title{Plot_SampleWise}
\usage{
Plot_SampleWise(df, value_col)
}
\arguments{
\item{df}{an adapted dataframe ready for analysis}

\item{value_col}{what value should be plotted}
}
\value{
a list of box and whiskers plots, with samples connected by lines across
their trajectories
}
\description{
The goal of this function is to generate a box and whisker plots with a connected
line graph showing the change in a value for an item over time.
}
