% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{Pr}
\alias{Pr}
\title{Pr}
\usage{
Pr(lc, lmin = 1, lmax, D)
}
\arguments{
\item{lc}{the variable this will be integrating over}

\item{lmin}{the lower bound of the distribution}

\item{lmax}{the upper bound of the distribution}

\item{D}{value of df}
}
\description{
The probability distribution used for the integration portion of D calculation
}
