% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMStoD.R
\name{SetSystemConfiguration}
\alias{SetSystemConfiguration}
\title{SetSystemConfiguration}
\usage{
SetSystemConfiguration(
  RIDefinitions,
  na_i,
  na_c,
  CenterLambda,
  IsImmersedInOil,
  IsCellGlassInterface
)
}
\arguments{
\item{RIDefinitions}{the collection of Refraction indices used refers to RIDefinition object in the matlab code}

\item{na_i}{Illumination NA}

\item{na_c}{Collection NA}

\item{CenterLambda}{the center wavelenth based on k}

\item{IsImmersedInOil}{True/False if so the immersion_ri will be set to the RI of the glass found in ri_def}

\item{IsCellGlassInterface}{whether there is a cell glass interface, if true then the fresnel coefficients will reflect
that the cell_glass interface is the reference plane. Otherwise the cell/media interface will be used}

\item{IsGlassCellInterface}{T/F Used for calculating the Fresnel Coefficient.}

\item{ImmersionRI}{The refractive index of the immersion media used for the objective}
}
\value{
a named list with the following items
RIDefinitions used for input
na_i: See params
na_c: see params
CenterLambda: see params
IsImmersedInOil: see params
IsCellGlassInterface: see params
DoF: Depth of focus
CenterWavelengthInVacuum: Used for D calculations
FresnelCoefficient: Used for calculating D (Need to learn more)
ImmersionRI: RI of immersion media
}
\description{
This system configuration object contains fields that are required to calculate
D. This instatiation method automatically calculates fields that are calculated from the inptus
}
\details{
A function that creates an class-lite kind of object as the original object in matlab
is basically just being used as a named list so there's no need to really overcomplicate things
here as I'm not really interested in refactoring code just creating something close to a 1to1 copy
}
