% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWSAnalysis.R
\name{t.testr}
\alias{t.testr}
\title{t.testr}
\usage{
\method{t}{testr}(df, value_col, paired = T)
}
\arguments{
\item{df}{an adapted dataframe ready for analysis}

\item{value_col}{what value should be plotted}

\item{paired}{should the t-test be paired or not.}
}
\value{
a list of box and whiskers plots, with samples connected by lines across
their trajectories
}
\description{
Perform t.tests between timepoints for each condition
}
